requirements.txt
dask==2021.12.0
matplotlib==3.4.3
multiprocess==0.70.12.2
numpy==1.21.4
pandas==1.3.4
pyserial==3.5
scipy==1.7.2
streamz==0.6.3
